backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.7.0
pydantic-settings>=2.1.0
pdfplumber==0.10.3
PyMuPDF==1.23.8
Pillow==10.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
google-generativeai==0.3.1
openai>=1.12.0
httpx>=0.27.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
pytesseract==0.3.10
imagehash==4.3.1
reportlab==4.0.7
python-docx==1.1.0
